@@ -1,10 +1,9 @@
-import { Module } from "@nestjs/common";
-import { CloudinaryProvider } from "./cloudinary.provider";
-import { CloudinaryService } from "./service/cloudinary.service";
+import { Module } from '@nestjs/common';
+import { CloudinaryProvider } from './cloudinary.provider';
+import { CloudinaryService } from './service/cloudinary.service';
 
 @Module({
-    providers: [CloudinaryProvider, CloudinaryService],
-    exports: [CloudinaryService],
+  providers: [CloudinaryProvider, CloudinaryService],
+  exports: [CloudinaryService],
 })
-
 export class CloudinaryModule {}
