@@ -1,7 +1,7 @@
 import { InjectModel } from '@nestjs/mongoose';
 import { Injectable } from '@nestjs/common';
 import { Product } from 'src/module/products/model/product.model';
-import { Model } from 'mongoose';
+import { DeleteResult, Model } from 'mongoose';
 
 @Injectable()
 export class ProductService {
@@ -43,7 +43,7 @@ export class ProductService {
     return [product];
   }
 
-  async deleteProduct(id: string) {
+  async deleteProduct(id: string): Promise<DeleteResult> {
     return this.productModel.deleteOne({ _id: id });
   }
 }
